removed logic from main class to test it
